Adds failcheck and changes h to double

+++ src/aiprog/model/Move.java
@@ -2,8 +2,8 @@ package aiprog.model;
 
 public class Move {
 	private Direction direction;
-	private int heuristic;
-	public Move(Direction direction, int heuristic){
+	private double heuristic;
+	public Move(Direction direction, double heuristic){
 		this.direction = direction;
 		this.heuristic = heuristic;
 	}
@@ -13,10 +13,10 @@ public class Move {
 	public void setDirection(Direction direction) {
 		this.direction = direction;
 	}
-	public int getHeuristic() {
+	public double getHeuristic() {
 		return heuristic;
 	}
-	public void setHeuristic(int heuristic) {
+	public void setHeuristic(double heuristic) {
 		this.heuristic = heuristic;
 	}
